+++ client-qa-support/src/main/java/org/realityforge/replicant/client/test/ReplicantClientTestModule.java
@@ -1,12 +1,13 @@
 package org.realityforge.replicant.client.test;
 
+import com.google.inject.Provides;
+import javax.inject.Singleton;
 import org.realityforge.guiceyloops.shared.AbstractModule;
 import org.realityforge.replicant.client.EntityChangeBroker;
-import org.realityforge.replicant.client.EntityChangeBrokerImpl;
 import org.realityforge.replicant.client.EntityRepository;
-import org.realityforge.replicant.client.EntityRepositoryImpl;
 import org.realityforge.replicant.client.EntitySubscriptionManager;
-import org.realityforge.replicant.client.EntitySubscriptionManagerImpl;
+import org.realityforge.replicant.client.EntitySystem;
+import org.realityforge.replicant.client.EntitySystemImpl;
 
 /**
  * Module containing all the common client services.
@@ -17,8 +18,27 @@ public final class ReplicantClientTestModule
   @Override
   protected void configure()
   {
-    bindSingleton( EntityRepository.class, EntityRepositoryImpl.class );
-    bindSingleton( EntityChangeBroker.class, EntityChangeBrokerImpl.class );
-    bindSingleton( EntitySubscriptionManager.class, EntitySubscriptionManagerImpl.class );
+    bind( EntitySystem.class ).to( EntitySystemImpl.class ).asEagerSingleton();
+  }
+
+  @Provides
+  @Singleton
+  public final EntityRepository getEntityRepository( final EntitySystem system )
+  {
+    return system.getRepository();
+  }
+
+  @Provides
+  @Singleton
+  public final EntityChangeBroker getEntityChangeBroker( final EntitySystem system )
+  {
+    return system.getChangeBroker();
+  }
+
+  @Provides
+  @Singleton
+  public final EntitySubscriptionManager getEntitySubscriptionManager( final EntitySystem system )
+  {
+    return system.getSubscriptionManager();
   }
 }